feat: add authentication and authorization

+++ src/services/notes/controllers/note-controller.js
@@ -1,17 +1,22 @@
 import { nanoid } from "nanoid";
 import notes from "../notes.js";
 import response from "../../../utils/response.js";
-import { ClientError, InvariantError, NotFoundError } from "../../../exceptions/index.js";
+import { AuthorizationError, InvariantError, NotFoundError } from "../../../exceptions/index.js";
 import noteRepositories from "../repositories/note-repositories.js";
 
 // Handler for NOTES
 const addNoteHandler = async (req, res, next) => {
   const { title, tags, body } = req.validated;
+  const owner = req.user.username;
+
+  console.log('req.user.id:', req.user.id);
+  console.log('owner:', owner);
 
   const note = await noteRepositories.createNote({
     title,
     body,
-    tags
+    tags,
+    owner
   });
 
   if (!note) {
@@ -23,10 +28,12 @@ const addNoteHandler = async (req, res, next) => {
 
 // GET ALL NOTES
 const getAllNotesHandler = async (req, res, next) => {
+  // Req user for auth
+  const owner = req.user;
   // Add req query
   const {title} = req.query;
 
-  const notes = await noteRepositories.getAllNotes();
+  const notes = await noteRepositories.getAllNotes(owner);
 
   if (title !== '') {
     const note = notes.filter((note) => note.title === title);
@@ -39,6 +46,15 @@ const getAllNotesHandler = async (req, res, next) => {
 
 const getNoteByIdHandler = async (req, res, next) => {
   const { id } = req.params;
+    // Req user for auth
+  const owner = req.user;
+
+  const isOwner = await noteRepositories.verifyNoteOwner(owner);
+
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses catatan ini'));
+  }
+
   const note = await noteRepositories.getNoteById(id);
 
   if (!note) {
@@ -52,6 +68,14 @@ const editNotesByIdHandler = async (req, res, next) => {
   const { id } = req.params;
 
   const {title, body, tags} = req.validated;
+
+  const owner = req.user;
+
+  const isOwner = await noteRepositories.verifyNoteOwner(owner);
+
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses catatan ini'));
+  }
 
   const updatedNote = await noteRepositories.editNoteById({
     id,
@@ -69,6 +93,15 @@ const editNotesByIdHandler = async (req, res, next) => {
 
 const deleteNotesByIdHandler = async (req, res, next) => {
   const { id } = req.params;
+
+  const owner = req.user;
+
+  const isOwner = await noteRepositories.verifyNoteOwner(owner);
+
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses catatan ini'));
+  }
+
   const deletedNote = await noteRepositories.deleteNoteById(id);
 
   if (!deletedNote) {
